+++ packages/widget/examples/landing-page/src/app/page.tsx
@@ -51,6 +51,7 @@ const consumerExamples = {
     },
     targetTokens: [ETH, USDC, USDT],
     targetChainIds: [42161, 43114],
+    protocolName: 'GMX',
   },
   hercules: {
     customTheme: {
@@ -62,21 +63,23 @@ const consumerExamples = {
     },
     targetTokens: [METISUSDC],
     targetChainIds: [1088],
+    protocolName: 'Hercules',
   },
   dfk: {
     customTheme: {
       bgColor: 'light',
-      '--synapse-text': 'hsl(12deg 85% 13%)',
-      '--synapse-secondary': 'hsl(12deg 85% 20%)',
+      '--synapse-text': 'hsl(12deg 85% 10%)',
+      '--synapse-secondary': 'hsl(12deg 40% 40%)',
       '--synapse-select-bg': 'hsl(35deg 100% 87%)',
       '--synapse-surface': 'hsl(32deg 69% 78%)',
       '--synapse-root': 'hsl(35deg 100% 87%)',
       '--synapse-border': 'hsl(29deg 53% 68%)',
-      '--synapse-focus': 'hsl(12deg 85% 15%)',
+      '--synapse-focus': 'hsl(12deg 85% 25%)',
       '--synapse-accent': 'hsl(12deg 85% 15%)',
     },
     targetTokens: [ETH, USDC, JEWEL, XJEWEL, WJEWEL],
     targetChainIds: [53935, 43114],
+    protocolName: 'DFK',
   },
 }
 
@@ -87,7 +90,7 @@ const customRpcs: CustomRpcs = {
 
 export default function Home() {
   const [config, setConfig] = useState<any>(initialConfig)
-  const [container, setContainer] = useState(true)
+  const [container, setContainer] = useState(false)
 
   const { web3Provider } = useEthereumWallet()
 
@@ -139,6 +142,7 @@ export default function Home() {
                 targetTokens={config.targetTokens}
                 targetChainIds={config.targetChainIds}
                 customTheme={config.customTheme}
+                protocolName={config.protocolName}
                 container={container}
               />
             </div>